fencing runtime rejection + lazy renewal; WAL segment-hole fix; SQL semantics regression tests
P1-2/P1-3 fencing: Branch::fence_gate called at all three write entries
(commit_tx/write_branch_commit/checkpoint_branch) inside commit_mu;
expired lease -> 40001; lazy renewal on commit path (<=1 PUT per ttl/3,
no background thread). Tests: multi_node.rs fence_expired_writer_rejected
+ fence_renew_keeps_healthy_writer_writing. Docs: fence.rs module comment,
multi-node design doc §3 status table rewritten to match implementation.

P0-1 finish: flush_now counters cleared-in-lock on success / returned on
failure; Bytes handle instead of full-segment memcpy; TRAILER_LEN constant.
Tests caught a real data-loss bug: cur_seg advanced on failed PUT leaves
segment-number holes, probe_tail contiguity assumption then drops every
later segment on recovery. Fixed: advance seg number only after successful
PUT (retry reuses path; superset overwrite under Uncertain).

WAL corruption regression suite (tests/wal_corruption.rs, 10 tests):
FrameIter bad len/CRC/truncated payload/truncated header/<=32B torn tail;
e2e open fail-fast on corrupted segment; flush PUT failure injection
(persistent-failure bounded error, transient self-heal).

SQL semantics regression (tests/sql_semantics.rs, 4 tests) caught two more
real bugs, fixed: failed_txn never cleared after ROLLBACK/COMMIT (session
permanently poisoned with 25P02); COMMIT in aborted transaction committed
partial writes (now discards, reports ROLLBACK, per PG).

S4 deviation recorded in SPEC 07 (integer widening to Int64, 22003).
clippy --workspace --all-targets zero; CI restored to -D warnings.
TASK.md restructured with evidence columns; round-2 response corrected
(false P0-3 claim and phantom test names marked as such).

184 tests pass, slt 7/7.

+++ crates/dendro-columnar/tests/footer_prune.rs
@@ -69,7 +69,7 @@ fn footer_metadata_and_pruning() {
         assert_eq!(rg.rows, 1000);
         assert_eq!(rg.first_row, (ri * 1000) as u64);
         let b = &rg.cols[0].blocks[0];
-        let lo = (ri as i64 * 1000) as i64;
+        let lo = ri as i64 * 1000;
         let hi = lo + 999;
         assert_eq!(b.min, ((lo as u64) ^ (1 << 63)), "rg{ri} pk min");
         assert_eq!(b.max, ((hi as u64) ^ (1 << 63)), "rg{ri} pk max");
